+++ src/vulkan/image_descriptor.cc
@@ -163,7 +163,7 @@ void ImageDescriptor::UpdateDescriptorSetIfNeeded(
   write.dstSet = descriptor_set;
   write.dstBinding = binding_;
   write.dstArrayElement = 0;
-  write.descriptorCount = image_infos.size();
+  write.descriptorCount = static_cast<uint32_t>(image_infos.size());
   write.descriptorType = GetVkDescriptorType();
   write.pImageInfo = image_infos.data();
